Refactor callback from example-app to cli generator
