--- example-apps/callback/.gitignore
@@ -1,6 +1,0 @@
-build
-docs
-node_modules
-*.log
-.environment
-.env
