--- example-apps/callback/.travis.yml
@@ -1,7 +1,0 @@
-language: node_js
-node_js:
-  - "12"
-before_script: 'npm install -g zapier-platform-cli'
-script: 'zapier test'
-notifications:
-  email: false
