--- example-apps/callback/README.md
@@ -1,7 +1,0 @@
-# "Callback" Example App For Zapier Platform
-
-[![Build Status](https://travis-ci.org/zapier/zapier-platform-example-app-callback.svg?branch=master)](https://travis-ci.org/zapier/zapier-platform-example-app-callback)
-
-A barebones app that has a callback defined.
-
-> We recommend using the zapier-platform-cli and `zapier init . --template=callback` to create an app.
