--- example-apps/callback/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "zapier-platform-example-app-callback",
-  "version": "1.0.0",
-  "description": "An example app for the Zapier platform.",
-  "repository": "zapier/zapier-platform-example-app-callback",
-  "homepage": "https://zapier.com/",
-  "author": "Bryan Helmig <bryan@zapier.com>",
-  "license": "BSD-3-Clause",
-  "main": "index.js",
-  "scripts": {
-    "test": "mocha --recursive"
-  },
-  "engines": {
-    "node": ">=8.10.0",
-    "npm": ">=5.6.0"
-  },
-  "dependencies": {
-    "zapier-platform-core": "10.1.1"
-  },
-  "devDependencies": {
-    "mocha": "^5.2.0",
-    "should": "^13.2.0"
-  },
-  "private": true
-}
