+++ packages/cli/src/generators/templates/callback/README.md
@@ -1,0 +1,5 @@
+# The "Callback" Template
+
+This example has a create showcasing the `performResume` callback function.
+
+Find out more in the docs: https://github.com/zapier/zapier-platform/blob/master/packages/cli#zgeneratecallbackurl.
