+++ packages/cli/src/generators/templates/callback/creates/prediction.js
