+++ packages/cli/src/generators/templates/callback/index.js
@@ -7,18 +7,6 @@ const App = {
   version: require('./package.json').version,
   platformVersion: require('zapier-platform-core').version,
 
-  beforeRequest: [],
-
-  afterResponse: [],
-
-  resources: {},
-
-  // If you want your trigger to show up, you better include it here!
-  triggers: {},
-
-  // If you want your searches to show up, you better include it here!
-  searches: {},
-
   // If you want your creates to show up, you better include it here!
   creates: {
     [prediction.key]: prediction,
